requirements.txt
# Core dependencies
chatterbox-tts==0.1.2
openai-whisper==20231117
openai==1.3.0
pyaudio==0.2.13
webrtcvad==2.0.10
sounddevice==0.4.6
numpy==1.26.0
scipy==1.11.4
python-dotenv==1.0.0
aiohttp==3.9.0
pynput==1.8.1  # For push-to-talk functionality
requests==2.31.0  # For LLM server health checks

# Audio processing
librosa==0.11.0
resampy==0.4.3
soundfile==0.12.1

# UI (optional)
gradio==4.19.0

# Utils
colorama==0.4.6
tqdm==4.66.1

# Google Gemini API
google-generativeai==0.5.0 